backend/app/requirements.txt
fastapi
uvicorn
pydantic[email]
motor
python-jose[cryptography]
python-dotenv
pydantic-settings
httpx
bcrypt
# (opzionale: puoi rimuovere del tutto "passlib[bcrypt]" se non lo usi altrove)
